backend/requirements.txt
dj-database-url==0.5.0
Django==2.2.3
psycopg2==2.8.3
whitenoise==4.1.3
django-extensions==2.1.9
Werkzeug==0.15.4
requests==2.22.0
django-debug-toolbar==2.0
pycpfcnpj==1.5.1
django-maintenance-mode==0.13.3
django-celery-email==2.0.2
ptipython==1.0.1
ptvsd==4.2.10

# Async
celery==4.3.0
django-celery-beat==1.5.0
redis==3.2.1
django-redis==4.10.0

# Channels
channels==2.2.0
channels-redis==2.4.0
daphne==2.3.0

# Custom app settings
django-appconf==1.0.3

# REST
djangorestframework==3.10.0
django-filter==2.1.0
Markdown==3.1.1
djangorestframework-csv==2.1.0
djangorestframework-msgpack==1.0.2
djangorestframework-xml==1.4.0
djangorestframework-yaml==1.0.3
rest-condition==1.0.3
djangorestframework-jwt==1.11.0
django-cors-headers==3.0.2

# Social
facebook-sdk==3.1.0

# Storages
django-storages==1.7.1
boto3==1.9.188

# Monitoring and error handling
scout-apm==2.1.0
sentry-sdk==0.10.2
django-prometheus==1.0.15

# Payments
six==1.12.0
python-dateutil==2.8.0
xmltodict==0.12.0
